--- jscomp/bin/js_compiler.mli
@@ -1,1 +1,0 @@
-(* *)
