+++ jscomp/snapshot.ninja
@@ -62,6 +62,10 @@ build $SNAP/unstable/bspack.ml: bspack | ./bin/bspack.exe
     flags = -bs-MD -module-alias Config=Config_whole_compiler  -I $OCAML_SRC_PARSING -I $OCAML_SRC_UTILS -I stubs -I ext -I common -I depends -I core -I main -bs-main Bspack_main
     main = Bspack_main
 
+build $SNAP/unstable/js_compiler.ml: bspack | ./bin/bspack.exe
+    flags = -D BS_COMPILER_IN_BROWSER=true -D BS_RELEASE_BUILD=true -bs-MD  -module-alias Config=Config_whole_compiler  -bs-exclude-I config   -I $OCAML_SRC_UTILS -I $OCAML_SRC_PARSING -I $OCAML_SRC_TYPING -I $OCAML_SRC_BYTECOMP -I $OCAML_SRC_DRIVER -I stubs -I ext -I syntax -I depends -I common -I core -I super_errors -I bsb -I outcome_printer -I main
+    main  = Jsoo_main
+# -o $@
 # Check it later 
 # rule bsbnative
 #     command = $ocamlopt -w -a  unix.cmxa str.cmxa  ./stubs/ext_basic_hash_stubs.c $in -o $out
